+++ pokedex/src/contexts/GlobalContext/GlobalStateContext.js
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export const GlobalContext = React.createContext()
